FUNCTIONAL INVENTORY SYSTEM
- HOLY SHIT THIS TOOK A WEEK OF NONSTOP WORK TO GET DONE FUCK YES -

*Press G to give the player a random item for their inventory*

Also added pit monster that crushes players with it's spiked tentacles.

Animations for this test boss are in blocking stage and by no means
done. You can find it in the 'Boos_Fight_Room_Test' scene. I cannot
spell.

I'm mostly just commiting this because the inventory system is FINALLY
working after a ton of stress and frustration.

Praise the sun! \o/

-C